+++ examples/custom_checkers/c9104_ModuleNameViolation.py
@@ -1,0 +1,2 @@
+"""Example of C9104 module-name-violation."""
+# The name of this file is not in snake_case, hence violates conventions for module names.