+++ CHANGELOG.md
@@ -42,7 +42,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Custom checkers:
 
-- `invalid-name-checker`: Provide beginner-friendly error messages when reporting variable names that violate PEP8 naming conventions.
+- `invalid-name-checker`: Provide beginner-friendly error messages when reporting variable names that violate Python naming conventions.
 
 Pylint checkers v2.16: